split snippet and fix name
